+++ scripts/mcp-tool-pack-baseline.json
@@ -14,14 +14,11 @@
   "contribute_to_hive",
   "create_backlog_item",
   "create_build_epic",
-  "create_digital_product",
   "create_epic",
   "create_portal_pr",
   "create_release_bundle",
   "deploy_feature",
   "diagnose_sandbox",
-  "dpf_test_kernel_refuse_probe",
-  "drive_browser_task",
   "edit_sandbox_file",
   "execute_promotion",
   "generate_code",
@@ -31,7 +28,6 @@
   "get_build_progress_visibility",
   "get_build_sandbox_state",
   "get_build_scoped_verification",
-  "get_finance_period_summary",
   "get_next_recommended_work",
   "get_release_status",
   "inspect_build_code_impact",
@@ -40,7 +36,6 @@
   "list_backlog_items",
   "list_build_activity_since",
   "list_epics",
-  "list_patch_posture",
   "list_sandbox_files",
   "principle_decide",
   "process_backlog_for_build_studio",
@@ -61,7 +56,6 @@
   "record_local_integration_result",
   "recover_sandbox",
   "register_digital_product_from_build",
-  "request_coworker",
   "retire_backlog_item",
   "reviewBuildPlan",
   "reviewDesignDoc",
@@ -78,16 +72,13 @@
   "schedule_release_bundle",
   "search_sandbox",
   "set_change_disposition",
-  "setup_email",
   "size_backlog_item",
   "start_build",
   "start_ideate_research",
   "start_sandbox",
   "start_scout_research",
   "submit_feedback",
-  "summon_coworker",
   "triage_backlog_item",
-  "trigger_contributor_inventory_sync",
   "update_backlog_item",
   "update_backlog_item_status",
   "update_epic",
